Creation of sqlite3 access functions module
The access functions for working with sqlite3 database were modularized.

It was created a folder named sqlite3_module into the folder of the main project.

In the main.py file the sqlite3_module was called with the namespace:

from sqlite3_module.sqlite_mod import *

In sqlite3_module was placed teh file sqlite_mod.py which contains sqlite3 functions.
Also, it was created __ini__.py file so python can create the namespace to the file.